packages/stock-counter-client/tests/unit-tests/defines/reports/salesreport.define.spec.ts
import { IsalesReport } from '@open-stock/stock-universal';
import Axios from 'axios-observable';
import { of } from 'rxjs';
import { beforeEach, describe, expect, expectTypeOf, it, vi } from 'vitest';
import { SalesReport } from '../../../../../stock-counter-client/src/defines/reports/salesreport.define';
import { StockCounterClient } from '../../../../../stock-counter-client/src/stock-counter-client';
import { createMockSalesReport, createMockSalesReports } from '../../../../../tests/stock-counter-mocks';

describe('Environment', () => {
  let instance: SalesReport;
  const axiosMock = { } as Axios;
  const companyId = 'companyid';

  beforeEach(() => {
    new StockCounterClient(axiosMock);
    instance = createMockSalesReport();
  });

  it('#StockCounterClient should have all callable static properties', () => {
    expect(StockCounterClient.calcCtrl).toBeDefined();
    expect(StockCounterClient.ehttp).toBeDefined();
    expect(StockCounterClient.logger).toBeDefined();
  });

  it('should have properties defined', () => {
    expect(instance).toBeInstanceOf(SalesReport);
    expect(instance._id).toBeDefined();
    expect(instance.createdAt).toBeDefined();
    expect(instance.updatedAt).toBeDefined();
    expect(instance.urId).toBeDefined();
    expect(instance.totalAmount).toBeDefined();
    expect(instance.date).toBeDefined();
    expect(instance.estimates).toBeDefined();
    expect(instance.invoiceRelateds).toBeDefined();
  });

  it('#getSalesReports static should get SalesReports array', async() => {
    const lSpy = vi.spyOn(StockCounterClient.ehttp, 'makeGet').mockImplementationOnce(() => of(createMockSalesReports(10)));
    const list = await SalesReport.getSalesReports(companyId, '/', 0, 0);

    expect(typeof list).toEqual('object');
    expectTypeOf(list).toEqualTypeOf<SalesReport[]>([]);
    expect(lSpy).toHaveBeenCalled();
  });

  it('#getOneSalesReport static should get one SalesReport', async() => {
    const lSpy = vi.spyOn(StockCounterClient.ehttp, 'makeGet').mockImplementationOnce(() => of(createMockSalesReport()));
    const one = await SalesReport.getOneSalesReport(companyId, 'urId');

    expect(typeof one).toEqual('object');
    expect(one).toBeInstanceOf(SalesReport);
    expect(lSpy).toHaveBeenCalled();
  });

  it('#addSalesReport static should add one SalesReport', async() => {
    const lSpy = vi.spyOn(StockCounterClient.ehttp, 'makePost').mockImplementationOnce(() => of({ success: true }));
    const added = await SalesReport.addSalesReport(companyId, createMockSalesReport() as IsalesReport);

    expect(typeof added).toEqual('object');
    expect(added).toHaveProperty('success');
    expect(added.success).toEqual(true);
    expect(typeof added.success).toBe('boolean');
    expectTypeOf(added.success).toEqualTypeOf<boolean>(Boolean('true'));
    expect(lSpy).toHaveBeenCalled();
  });

  it('#deleteSalesReports static should delete many SalesReports', async() => {
    const lSpy = vi.spyOn(StockCounterClient.ehttp, 'makePut').mockImplementationOnce(() => of({ success: true }));
    const deleted = await SalesReport.deleteSalesReports(companyId, ['_ids']);

    expect(typeof deleted).toEqual('object');
    expect(deleted).toHaveProperty('success');
    expect(deleted.success).toEqual(true);
    expect(typeof deleted.success).toBe('boolean');
    expectTypeOf(deleted.success).toEqualTypeOf<boolean>(Boolean('true'));
    expect(lSpy).toHaveBeenCalled();
  });
});
